requirements.txt
# Install the required packages
smbus2
paho-mqtt
requests
matplotlib
numpy
scipy
RPi.GPIO

